move compat funcs to utils, no circ imports that way
